pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "agentflow"
version = "0.1.0"
description = "Parallel and DAG orchestration for Codex, Claude, and Kimi agents"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "OpenAI Codex CLI" }]
dependencies = [
  "boto3>=1.35.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "pydantic>=2.11.0",
  "pyyaml>=6.0.2",
  "typer>=0.16.0",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
]

[project.scripts]
agentflow = "agentflow.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["agentflow"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

